31

Approved-by: Przemysław Seń <sen.przemyslaw@gmail.com>

+++ src/main/java/com/blinenterprise/SyropKlonowy/repository/ProductRepository.java
@@ -5,6 +5,7 @@ import com.blinenterprise.SyropKlonowy.domain.Product;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -16,11 +17,11 @@ public interface ProductRepository extends CrudRepository<Product, Long> {
 
     List<Product> findAllByCategory(Category category);
 
-    List<Product> findAllByPriceLessThanEqual(Double price);
+    List<Product> findAllByPriceLessThanEqual(BigDecimal price);
 
-    List<Product> findAllByPriceGreaterThanEqual(Double price);
+    List<Product> findAllByPriceGreaterThanEqual(BigDecimal price);
 
-    List<Product> findAllByPriceLessThanEqualAndCategory(Double price, Category category);
+    List<Product> findAllByPriceLessThanEqualAndCategory(BigDecimal price, Category category);
 
-    List<Product> findAllByPriceGreaterThanEqualAndCategory(Double price, Category category);
+    List<Product> findAllByPriceGreaterThanEqualAndCategory(BigDecimal price, Category category);
 }
